Add repository implementation

+++ src/main/java/net/lxwrz/simplees/infrastructure/EventStore.java
@@ -4,6 +4,6 @@ import java.util.List;
 import java.util.UUID;
 
 public interface EventStore {
-    void saveEvents(UUID aggregateId, List<Event> events, int expectedVersion) throws VersionConflict;
+    void saveEvents(UUID aggregateId, List<Event> events, long expectedVersion) throws VersionConflict;
     List<Event> loadEvents(UUID aggregateId) throws EventStreamNotFound;
 }
